requirement.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.5
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
dateutils==0.6.8
Django==3.0.6
fake-useragent==0.1.11
future==0.18.2
googlefinance==0.7
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.1
ipynb-py-convert==0.4.5
kombu==4.6.10
lxml==4.5.1
multitasking==0.0.9
newsapi-python==0.2.6
nsepy==0.8
nsetools==1.0.11
numpy==1.18.5
pandas==1.0.4
pandas-datareader==0.8.1
parse==1.15.0
psycopg2==2.7.5
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
razorpay==1.2.0
redis==3.5.3
requests==2.23.0
requests-html==0.10.0
simplejson==3.17.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.46.1
urllib3==1.25.9
vine==1.3.0
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
yahoo-fin==0.8.5
yahoo-finance==1.4.0
#yahoo-finance-api==0.0.1
yfinance==0.1.54
zipp==3.1.0
channels-redis==2.4.2
